bakend_part/requirements-dev.txt
absl-py==2.1.0
altgraph==0.17.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.3.0
asgiref==3.8.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
Authlib==1.3.0
bcrypt==4.3.0
billiard==4.2.1
cachetools==5.3.2
celery==5.5.3
certifi==2023.11.17
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.9.2
cron-descriptor==1.4.5
cryptography==42.0.7
distlib==0.3.8
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.2
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-mssql-backend==2.8.1
django-pyodbc-azure==2.1.0.0
django-timezone-field==7.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
dnspython==2.6.1
dotenv==0.9.9
drf-spectacular==0.28.0
ecdsa==0.19.1
email_validator==2.1.1
environ==1.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.13.1
flatbuffers==25.2.10
gast==0.4.0
gevent==25.8.2
google-auth==2.27.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.60.0
h11==0.14.0
h5py==3.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
identify==2.6.13
idna==3.6
inflection==0.5.1
iniconfig==2.1.0
Jinja2==3.1.4
jsonschema==4.24.1
jsonschema-specifications==2025.4.1
keras==3.10.0
kombu==5.5.4
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
ngrok==1.4.0
nodeenv==1.9.1
npm==0.1.1
numpy==1.26.3
oauthlib==3.2.2
opencv-python==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
optional-django==0.1.0
optree==0.17.0
orjson==3.10.3
packaging==23.2
pandas==2.2.2
passlib==1.7.4
pefile==2023.2.7
pillow==10.2.0
platformdirs==4.1.0
plotly==5.22.0
pluggy==1.6.0
pre_commit==4.3.0
prompt_toolkit==3.0.51
protobuf==5.29.5
psycopg2-binary==2.9.10
pyarrow==15.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.7.1
pydantic-settings==2.9.1
pydantic_core==2.18.2
Pygments==2.18.0
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
PyJWT==2.9.0
pyodbc==5.1.0
pypng==0.20220715.0
pytest==8.4.0
pytest-asyncio==1.0.0
pytest-django==4.11.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
python-json-logger==3.3.0
python-multipart==0.0.9
pytz==2024.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
qrcode==7.4.2
redis==5.2.1
referencing==0.36.2
requests==2.31.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
rich==13.7.1
rpds-py==0.26.0
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.4.4
starlette==0.37.2
tenacity==8.3.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.19.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tomli==2.2.1
typer==0.12.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
uritemplate==4.2.0
urllib3==2.1.0
uvicorn==0.29.0
vine==5.1.0
virtualenv==20.25.0
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.1
wrapt==1.14.1
writer==0.6.0rc3
writer-sdk==0.1.0a2
zope.event==5.1.1
zope.interface==7.2
